Feature/read workbook 나만의 문제집 마이 페이지 조회 구현

+++ src/main/java/com/example/stepbackend/aggregate/dto/workbook/SelectedItemsDTO.java
@@ -9,5 +9,5 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class SelectedItemsDTO {
-    private List<Long> selectedItems;
+    private List<Integer> selectedItems;
 }
